Parse JSON parses nested arrays correctly
It was getting messed up previously because parseNumber() throws away
the last value that is considered. Therefore if it hit a ']', the
numbers loop would stop as expected, but then it would throw out the
']'. As a result parseArray would never see the ']' value and would
never end the array.

I included a variable called endOfArray which indicates when an array
has been ended inside the while loop of another parsing method.